Additions build fix.
svn:sync-xref-src-repo-rev: r121261

+++ src/VBox/Additions/common/VBoxService/VBoxServiceControlSession.cpp
@@ -1,4 +1,4 @@
-/* $Id: VBoxServiceControlSession.cpp 71314 2018-03-13 15:51:14Z andreas.loeffler@oracle.com $ */
+/* $Id: VBoxServiceControlSession.cpp 71316 2018-03-13 17:13:32Z andreas.loeffler@oracle.com $ */
 /** @file
  * VBoxServiceControlSession - Guest session handling. Also handles the spawned session processes.
  */
@@ -639,7 +639,7 @@ static int vgsvcGstCtrlSessionHandlePathUserDocuments(PVBOXSERVICECTRLSESSION pS
 
     /* Report back in any case. */
     int rc2 = VbglR3GuestCtrlMsgReplyEx(pHostCtx, rc, 0 /* Type */,
-                                        szPath, strlen(szPath) + 1 /* Include terminating zero */);
+                                        szPath, (uint32_t)strlen(szPath) + 1 /* Include terminating zero */);
     if (RT_FAILURE(rc2))
         VGSvcError("Failed to report user documents, rc=%Rrc\n", rc2);
     if (RT_SUCCESS(rc))
@@ -662,7 +662,7 @@ static int vgsvcGstCtrlSessionHandlePathUserHome(PVBOXSERVICECTRLSESSION pSessio
 
     /* Report back in any case. */
     int rc2 = VbglR3GuestCtrlMsgReplyEx(pHostCtx, rc, 0 /* Type */,
-                                        szPath, strlen(szPath) + 1 /* Include terminating zero */);
+                                        szPath, (uint32_t)strlen(szPath) + 1 /* Include terminating zero */);
     if (RT_FAILURE(rc2))
         VGSvcError("Failed to report user home, rc=%Rrc\n", rc2);
     if (RT_SUCCESS(rc))
